feat: ReplyMapper 구현 및 변수명 수정

+++ src/main/java/org/crud_ex/mapper/BoardMapper.java
@@ -9,18 +9,18 @@ public interface BoardMapper {
 	// 목록 조회
 	List<Board> findAll();
 
-	// 단건 조회
+	// 한 건 조회
 	Board findOne(Long boardId);
 
 	// 등록
-	Long save(Board board);
+	int save(Board board);
 
 	// 수정
-	Long update(Board board);
+	int update(Board board);
 
 	// 삭제
-	Long delete(Long boardId);
+	int delete(Long boardId);
 
 	// 총 개수 (페이징 용)
-	Long getTotalCount();
+	int getTotalCount();
 }
